initial push - small steps
HackSB Sustainability Category Winner 2025 - Small Steps is a website aimed to educate users about simple daily tasks that can be completed to be more mindful of the environment. It's a point-and-click story game which follows a character Daisy along with her leaf companion to teach the user about sustainability. There are two interactive clickable areas anchored to a minigame which demonstrates a simple task, such as turning off a running tap or cleaning up garbage, along with a text box displaying information about environmental issues and how we can make a difference to help.