initial push - PWA Trie Project
CS Academy Data Structures Trie Project - A project to practice utilizing Tries through a predictive writing assistant (PWA). Themed around Rosie, Rosé's album. There are two modes: typing using only the lyrics from the album along with autofill and probabilities on which character/word is next, and a "finish the lyric" game which prompts the user to write a word to complete a randomly selected lyric out of the album, along with hints.